Bunch of bug fixes for 500 internal service error

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/HelperDao.java
@@ -8,7 +8,7 @@ public interface HelperDao {
 
     Transfer transferFixer(Transfer transfer);
 
-    Transfer findTransferById(Integer id);
+    Transfer findTransferById(Long id);
 
     User findByAccountId(Long userID) throws UsernameNotFoundException;
